調整しろ

+++ ソースコード/engine/base/RandCreate.h
@@ -49,14 +49,14 @@ public:
 		//intRand(i, j);
 		return intRand(mt);
 	}
-	static double sGetRandFloat(double i, double j)
+	static float sGetRandFloat(double i, double j)
 	{
 		std::mt19937_64 mt;
 		std::random_device rnd;
 		mt.seed(rnd());
 		mt.seed(rnd());
 		std::uniform_real_distribution<> floatRand(i, j);
-		return floatRand(mt);
+		return (float)floatRand(mt);
 	}
 private:
